+++ engine/src/main/java/dev/gamekit/ui/Bounds.java
@@ -41,6 +41,14 @@ public class Bounds {
     this.height = height;
   }
 
+  public void setX(int x) { this.x = x; }
+
+  public void setY(int y) { this.y = y; }
+
+  public void setWidth(int width) { this.width = width; }
+
+  public void setHeight(int height) { this.height = height; }
+
   /** Copies values from another bounds object */
   public void set(Bounds bounds) {
     set(bounds.x, bounds.y, bounds.width, bounds.height);